Moving proof impl. in algorithm.rs
